Optimized rendering of the example
Rendering a lot of similar elements using the RN Skia Reconciler lacks a bit performance (a few PRs are waiting that will help).

The reconciler repeats the same steps with each element without optimizing this in any way - so changing the rendering of each confetti item from a <Group><RRect /></Group> to a <Drawing /> with custom drawing function will help.

Also remember that each computed value is evaluated on every frame if it has changed - so this will also add to the overhead here.

This PR updated the example to use a Drawing component with a custom drawing function instead of the <Group> / <RRect> components.

Each Confetti-component can then calculate its origin and matrix directly when rendering - without having to go through a computed value.

In addition a main animation value is used for driving all the confetti animations so that we don't need to run the animation for each confetti item.

This might be a bit less boilerplate than just using the built-in primitives and functionalities when rendering, but it also shows some of the pitfalls and their solutions when rendering a lot of items on screen at the same time.

It would even be possible to render all confetti items in one Drawing component, but that looked too boring :)